Fixing some issues && critical bug fix. (#3)
* Kaesereibe

* Logs in Folders

* Fix some Importing issues.

* Fix SearchEngine

* Critical Database bug fix.

* Implementing pisse.html

+++ Pisstaube/Controllers/PrivateAPIController.cs
@@ -2,8 +2,6 @@ using System;
 using System.Diagnostics.CodeAnalysis;
 using System.IO;
 using System.Linq;
-using System.Runtime.Serialization.Formatters.Binary;
-using System.Threading;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using osu.Framework.Logging;
@@ -28,16 +26,36 @@ namespace Pisstaube.Controllers
     [ApiController]
     public class PrivateAPIController : ControllerBase
     {
+        private readonly PisstaubeDbContextFactory _contextFactory;
+        private readonly Storage _storage;
+        private readonly BeatmapSearchEngine _searchEngine;
+        private readonly Crawler _crawler;
+        private readonly Cleaner _cleaner;
+        private readonly PisstaubeCacheDbContextFactory _cache;
+
         private static readonly object _lock = new object();
-        // GET /api/pisstaube/dump
-        [HttpGet("dump")]
-        public ActionResult DumpDatabase(
-            [FromServices] PisstaubeDbContext db,
-            [FromServices] Storage storage
-            )
+
+        public PrivateAPIController(PisstaubeDbContextFactory contextFactory,
+            Storage storage, BeatmapSearchEngine searchEngine, Crawler crawler, Cleaner cleaner,
+            PisstaubeCacheDbContextFactory cache)
         {
+            _contextFactory = contextFactory;
+            _storage = storage;
+            _searchEngine = searchEngine;
+            _crawler = crawler;
+            _cleaner = cleaner;
+            _cache = cache;
+        }
+        
+        // GET /api/pisstaube/dump?key={KEY}
+        [HttpGet("dump")]
+        public ActionResult DumpDatabase(string key)
+        {
+            if (key != Environment.GetEnvironmentVariable("PRIVATE_API_KEY"))
+                return Unauthorized("Key is wrong!");
+
             lock (_lock) {
-                var tmpStorage = storage.GetStorageForDirectory("tmp");
+                var tmpStorage = _storage.GetStorageForDirectory("tmp");
                 
                 if (tmpStorage.Exists("dump.piss"))
                     System.IO.File.Delete(tmpStorage.GetFullPath("dump.piss"));
@@ -45,10 +63,10 @@ namespace Pisstaube.Controllers
                 using (var dumpStream = tmpStorage.GetStream("dump.piss", FileAccess.Write))
                 using (var sw = new MStreamWriter(dumpStream))
                 {
-                    sw.Write(db.BeatmapSet.Count());
-                    foreach (var bmSet in db.BeatmapSet)
+                    sw.Write(_contextFactory.Get().BeatmapSet.Count());
+                    foreach (var bmSet in _contextFactory.Get().BeatmapSet)
                     {
-                        bmSet.ChildrenBeatmaps = db.Beatmaps.Where(bm => bm.ParentSetId == bmSet.SetId).ToList();
+                        bmSet.ChildrenBeatmaps = _contextFactory.Get().Beatmaps.Where(bm => bm.ParentSetId == bmSet.SetId).ToList();
                         sw.Write(bmSet);
                     }
                 }
@@ -61,8 +79,6 @@ namespace Pisstaube.Controllers
         // GET /api/pisstaube/put?key={}
         [HttpPut("put")]
         public ActionResult PutDatabase(
-            [FromServices] PisstaubeDbContext db,
-            [FromServices] BeatmapSearchEngine searchEngine,
             [FromQuery] string key,
             [FromQuery] bool drop
         )
@@ -72,13 +88,16 @@ namespace Pisstaube.Controllers
 
             if (drop)
             {
-                searchEngine.DeleteAllBeatmaps();
-                db.Database.ExecuteSqlCommand("SET FOREIGN_KEY_CHECKS = 0;" +
+                _searchEngine.DeleteAllBeatmaps();
+                using (var db = _contextFactory.GetForWrite()) {
+                    db.Context.Database.ExecuteSqlCommand
+                                             ("SET FOREIGN_KEY_CHECKS = 0;" +
                                               "TRUNCATE TABLE `Beatmaps`;" +
                                               "ALTER TABLE `Beatmaps` AUTO_INCREMENT = 1;" +
                                               "TRUNCATE TABLE `BeatmapSet`;" +
                                               "ALTER TABLE `BeatmapSet` AUTO_INCREMENT = 1;" +
                                               "SET FOREIGN_KEY_CHECKS = 1;");
+                }
             }
 
             lock (_lock) {
@@ -86,6 +105,7 @@ namespace Pisstaube.Controllers
                 
                 using (var stream = f.OpenReadStream())
                 using (var sr = new MStreamReader(stream))
+                using (var db = _contextFactory.GetForWrite())
                 {
                     var count = sr.ReadInt32();
                     Logger.LogPrint($"Count: {count}");
@@ -95,38 +115,17 @@ namespace Pisstaube.Controllers
                         var set = sr.ReadData<BeatmapSet>();
                         
                         Logger.LogPrint($"Importing BeatmapSet {set.SetId} {set.Artist} - {set.Title} ({set.Creator}) of Index {i}", LoggingTarget.Database, LogLevel.Important);
-                        if (db.BeatmapSet.Any(s => s.SetId == set.SetId)) {
-                            db.BeatmapSet.Update(set);
-                        } else {
-                            db.BeatmapSet.Add(set);
-                        }
-                    }
-                    
-                    db.SaveChanges();
-                    
-                    /*
-                    
-                    var b = new byte[4];
-                    
-                    stream.Read(b);
-                    var setCount = BitConverter.ToInt32(b);
 
-                    var bf = new BinaryFormatter();
-                    for (var i = 0; i < setCount; i++)
-                    {
-                        var set = (BeatmapSet) bf.Deserialize(stream);
-                        
-                        Logger.LogPrint($"Importing BeatmapSet {set.SetId} {set.Artist} - {set.Title} ({set.Creator}) of Index {i}", LoggingTarget.Database, LogLevel.Important);
-                        if (db.BeatmapSet.Any(s => s.SetId == set.SetId)) {
-                            db.BeatmapSet.Update(set);
-                        } else {
-                            db.BeatmapSet.Add(set);
-                        }
-                        searchEngine.IndexBeatmap(set);
-                        db.SaveChanges();
+                        if (!drop)
+                            if (db.Context.BeatmapSet.Any(s => s.SetId == set.SetId))
+                                db.Context.BeatmapSet.Update(set);
+                            else
+                                db.Context.BeatmapSet.Add(set);
+                        else
+                            db.Context.BeatmapSet.Add(set);
+                       
                     }
-                    
-                    */
+                    Logger.LogPrint("Finish importing maps!");
                 }
                 
                 return Ok("Success!");
@@ -147,32 +146,25 @@ namespace Pisstaube.Controllers
         private int cInt;
         // GET /api/pisstaube/stats
         [HttpGet("stats")]
-        public ActionResult GetPisstaubeStats(
-            [FromServices] Crawler crawler,
-            [FromServices] Cleaner cleaner,
-            [FromServices] PisstaubeDbContext context
-        )
+        public ActionResult GetPisstaubeStats()
         {
-            if (!crawler.IsCrawling && cInt == 0)
-                cInt = context.BeatmapSet.LastOrDefault()?.SetId + 1 ?? 0;
-            
+            lock (_lock)
+                if (!_crawler.IsCrawling && cInt == 0)
+                    cInt = _contextFactory.Get().BeatmapSet.LastOrDefault()?.SetId + 1 ?? 0;
+
             return Ok(new PisstaubeStats
             {
-                IsCrawling = crawler.IsCrawling,
-                LatestCrawledId = crawler.IsCrawling ? crawler.LatestId : cInt,
-                MaxStorage = cleaner.MaxSize,
-                StorageUsed = (ulong) cleaner.DataDirectorySize,
-                StorageUsagePercent = MathF.Round ((ulong)cleaner.DataDirectorySize / cleaner.MaxSize * 100, 2)
+                IsCrawling = _crawler.IsCrawling,
+                LatestCrawledId = _crawler.IsCrawling ? _crawler.LatestId : cInt,
+                MaxStorage = _cleaner.MaxSize,
+                StorageUsed = (ulong) _cleaner.DataDirectorySize,
+                StorageUsagePercent = MathF.Round ((ulong)_cleaner.DataDirectorySize / _cleaner.MaxSize * 100, 2)
             });
         }
 
 
         [HttpGet("recovery")]
         public ActionResult Recovery(
-            [FromServices] PisstaubeDbContext db,
-            [FromServices] BeatmapSearchEngine searchEngine,
-            [FromServices] Crawler crawler,
-            [FromServices] PisstaubeCacheDbContextFactory _cache,
             [FromQuery] string key,
             [FromQuery] RecoveryAction action
             )
@@ -183,59 +175,53 @@ namespace Pisstaube.Controllers
             switch (action)
             {
                 case RecoveryAction.RepairElastic:
-                    new Thread(() =>
-                    {
-                        Logger.LogPrint("Repairing ElasticSearch");
-                        crawler.Stop();
+                    Logger.LogPrint("Repairing ElasticSearch");
+                    _crawler.Stop();
 
-                        searchEngine.DeleteAllBeatmaps();
+                    _searchEngine.DeleteAllBeatmaps();
                     
-                        foreach (var beatmapSet in db.BeatmapSet)
-                        {
-                            beatmapSet.ChildrenBeatmaps = db.Beatmaps.Where(b => b.ParentSetId == beatmapSet.SetId).ToList();
-                            searchEngine.IndexBeatmap(beatmapSet);
-                        }
+                    foreach (var beatmapSet in _contextFactory.Get().BeatmapSet)
+                    {
+                        beatmapSet.ChildrenBeatmaps = _contextFactory.Get().Beatmaps.Where(b => b.ParentSetId == beatmapSet.SetId).ToList();
+                        _searchEngine.IndexBeatmap(beatmapSet);
+                    }
                     
-                        if (Environment.GetEnvironmentVariable("CRAWLER_DISABLED") != "true")
-                            crawler.BeginCrawling();
-                    }).Start();
+                    if (Environment.GetEnvironmentVariable("CRAWLER_DISABLED") != "true")
+                        _crawler.BeginCrawling();
                     break;
                 case RecoveryAction.RecrawlEverything:
-                    new Thread(() =>
-                    {
-                        Logger.LogPrint("Recrawl Everything!");
+                    Logger.LogPrint("Recrawl Everything!");
                     
-                        crawler.Stop();
-                        searchEngine.DeleteAllBeatmaps();
-                        db.Database.ExecuteSqlCommand("SET FOREIGN_KEY_CHECKS = 0;" +
+                    _crawler.Stop();
+                    _searchEngine.DeleteAllBeatmaps();
+                    using (var db = _contextFactory.GetForWrite()) {
+                        db.Context.Database.ExecuteSqlCommand("SET FOREIGN_KEY_CHECKS = 0;" +
                                                       "TRUNCATE TABLE `Beatmaps`;" +
                                                       "ALTER TABLE `Beatmaps` AUTO_INCREMENT = 1;" +
                                                       "TRUNCATE TABLE `BeatmapSet`;" +
                                                       "ALTER TABLE `BeatmapSet` AUTO_INCREMENT = 1;" +
                                                       "SET FOREIGN_KEY_CHECKS = 1;");
-
-                        using (var cacheDb = _cache.GetForWrite())
-                        {
-                            cacheDb.Context.Database.ExecuteSqlCommand(
-                                "DELETE FROM `CacheBeatmaps`;" +
-                                "DELETE FROM `CacheBeatmapSet`;");
-                        }
-                        crawler.BeginCrawling();
-                    }).Start();
+                    }
+                    using (var cacheDb = _cache.GetForWrite())
+                    {
+                        cacheDb.Context.Database.ExecuteSqlCommand(
+                            "DELETE FROM `CacheBeatmaps`;" +
+                            "DELETE FROM `CacheBeatmapSet`;");
+                    }
+                    _crawler.BeginCrawling();
                     break;
                 case RecoveryAction.RecrawlUnknown:
-                    new Thread(() =>
-                    {
-                        Logger.LogPrint("Recrawl All unknown maps!");
+                    Logger.LogPrint("Recrawl All unknown maps!");
                     
-                        crawler.Stop();
-                        for (var i = 0; i < db.BeatmapSet.Last().SetId; i++)
+                    _crawler.Stop();
+                    using (var db = _contextFactory.GetForWrite()) {
+                        for (var i = 0; i < db.Context.BeatmapSet.Last().SetId; i++)
                         {
-                            if (!db.BeatmapSet.Any(set => set.SetId == i))
-                                crawler.Crawl(i, db);
+                            if (!db.Context.BeatmapSet.Any(set => set.SetId == i))
+                                _crawler.Crawl(i, db.Context);
                         }
-                        crawler.BeginCrawling();
-                    }).Start();
+                    }
+                    _crawler.BeginCrawling();
                     break;
                 default:
                     return BadRequest("Unknown Action type!");
